requirements.txt
Flask==2.2.2
pandas==2.1.1
numpy==1.26.0
gunicorn==20.0.4
Werkzeug==2.2.3
python-dotenv=0.21.0
